solved errors

+++ src/main/java/com/genc/hrms/repository/PayrollRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface PayrollRepository extends JpaRepository<Payroll, Integer> {
+public interface PayrollRepository extends JpaRepository<Payroll, Long> {
 
     Payroll findTopByEmployee_EmployeeIdOrderByPayrollIdDesc(Long employeeId);
